refactor: moved errors logic
