+++ src/Contexts/shared/domain/errors/custom.error.ts
